ml-service/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
Pillow
numpy
